- feat(tests/makefile): Added comprehensive unit & integration tests for current components + Enhanced Makefile (#32)
* chore(build): Add Makefile and refactor build.sh to use it

* chore(build): Prompt user for compilation in build script

* test: add unit test for PING and ECHO commands

* test: isolate main() in server for clean test builds

* - feat(tests): Added comprehensive unit and integration tests for the current MemoraDB components

* - fix: Fixed the LRANGE assertion to handle multi-buffered TCP packts correctly + minor code comments cleanup.

---------

+++ src/parser/parser.c
@@ -117,7 +117,7 @@ void dispatch_command(int client_fd, char * tokens[], int token_count){
                 break;
             }
 
-            int total_elements = 0;
+            size_t total_elements = 0;
             for (int i = 2; i < token_count; i++) {
                 size_t new_len = list_rpush(list, tokens[i]);
                 if (new_len > total_elements) {
@@ -125,7 +125,7 @@ void dispatch_command(int client_fd, char * tokens[], int token_count){
                 }
             }
 
-            dprintf(client_fd, ":%d\r\n", total_elements);
+            dprintf(client_fd, ":%zu\r\n", total_elements);
         }
         break;
     case CMD_LPUSH:
@@ -138,7 +138,7 @@ void dispatch_command(int client_fd, char * tokens[], int token_count){
                 break;
             }
 
-            int total_elements = 0;
+            size_t total_elements = 0;
             for (int i = 2 ; i < token_count ; i++) {
                 size_t new_len = list_lpush(list, tokens[i]);
                 if (new_len > total_elements) {
@@ -146,7 +146,7 @@ void dispatch_command(int client_fd, char * tokens[], int token_count){
                 }
             }
 
-            dprintf(client_fd, ":%d\r\n", total_elements);
+            dprintf(client_fd, ":%zu\r\n", total_elements);
         }
         break;
     case CMD_LRANGE:
